Moved InvokeExtensions, CollectorExtensions and LambdaAdapters classes to bridge module
