+++ jvm-bridge/src/test/kotlin/com/kamelia/sprinkler/bridge/ExtensionsTest.kt
@@ -1,49 +1,10 @@
-package com.kamelia.sprinkler.util.jvmlambda
+package com.kamelia.sprinkler.bridge
 
-import java.util.concurrent.Callable
-import java.util.function.BiConsumer
-import java.util.function.BiFunction
-import java.util.function.BiPredicate
-import java.util.function.Consumer
-import java.util.function.DoubleBinaryOperator
-import java.util.function.DoubleConsumer
-import java.util.function.DoubleFunction
-import java.util.function.DoublePredicate
-import java.util.function.DoubleSupplier
-import java.util.function.DoubleToIntFunction
-import java.util.function.DoubleToLongFunction
-import java.util.function.DoubleUnaryOperator
-import java.util.function.Function
-import java.util.function.IntBinaryOperator
-import java.util.function.IntConsumer
-import java.util.function.IntFunction
-import java.util.function.IntPredicate
-import java.util.function.IntSupplier
-import java.util.function.IntToDoubleFunction
-import java.util.function.IntToLongFunction
-import java.util.function.IntUnaryOperator
-import java.util.function.LongBinaryOperator
-import java.util.function.LongConsumer
-import java.util.function.LongFunction
-import java.util.function.LongPredicate
-import java.util.function.LongSupplier
-import java.util.function.LongToDoubleFunction
-import java.util.function.LongToIntFunction
-import java.util.function.LongUnaryOperator
-import java.util.function.ObjDoubleConsumer
-import java.util.function.ObjIntConsumer
-import java.util.function.ObjLongConsumer
-import java.util.function.Predicate
-import java.util.function.Supplier
-import java.util.function.ToDoubleBiFunction
-import java.util.function.ToDoubleFunction
-import java.util.function.ToIntBiFunction
-import java.util.function.ToIntFunction
-import java.util.function.ToLongBiFunction
-import java.util.function.ToLongFunction
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
+import java.util.concurrent.Callable
+import java.util.function.*
 
 class ExtensionsTest {
