+++ utils/README.md
@@ -7,12 +7,7 @@
 - [Intentions](#intentions)
 - [CloseableScope](#closeablescope)
 - [Box Delegate](#box-delegate)
-- [Collector Utilities](#collector-utilities)
-  - [Collector Shorthands](#collector-shorthands)
-  - [Collector Factories](#collector-factories)
-- [Kotlin Lambda Adapters for Java](#kotlin-lambda-adapters-for-java)
-  - [LambdaAdapters](#lambdaadapters)
-  - [InvokeExtensions](#invokeextensions)
+- [Collector Factories](#collector-factories)
 - [ByteArrayDecoding](#bytearraydecoding)
 - [ByteAccess](#byteaccess)
 - [unsafeCast](#unsafecast)
@@ -34,8 +29,12 @@ But one might see that they are not entirely symmetrical in their usage as soon 
 
 ```java
 // java
-try (var resource1 = new Resource1(); var resource2 = new Resource2()) {
-    // Use resources
+class Main {
+    public static void main(String[] args) {
+        try (var resource1 = new Resource1(); var resource2 = new Resource2()) {
+            // Use resources
+        }
+    }
 }
 ```
 
@@ -54,11 +53,15 @@ Another issue arises in Java's case when we want to use a new resource within th
 already opened ones:
 
 ```java
-try (var resource1 = new Resource1()) {
-    resource1.doSomething();
-    // Use resource1
-    try (var resource2 = new Resource2()) {
-        // Use resource2
+class Main {
+    public static void main(String[] args) {
+        try (var resource1 = new Resource1()) {
+            resource1.doSomething();
+            // Use resource1
+            try (var resource2 = new Resource2()) {
+                // Use resource2
+            }
+        }
     }
 }
 ```
@@ -171,24 +174,7 @@ fun main() {
 }
 ```
 
-## Collector Utilities
-
-Sprinkler-utils brings a few utilities to simplify the creation of Java `Collector`s, and their usage which is sometimes
-a bit clunky.
-
-### Collector Shorthands
-
-Calling the different functional interfaces composing a `Collector` is often redundant : you obtain the element thanks
-to a getter, and then call the interface's method with the arguments. Sprinkler-utils provides a few shorthands to
-simplify this. They are all extension functions on `Collector` and are all inlined.
-
-- `supply` is a shorthand method for `Collector::supplier::get`
-- `accumulate` is a shorthand method for `Collector::accumulator::accept`
-- `combine` is a shorthand method for `Collector::combiner::apply`
-- `finish` is a shorthand method for `Collector::finisher::apply`
-- `characteristics` is a shorthand property for `Collector::characteristics`
-
-### Collector Factories
+## Collector Factories
 
 Java's standard library is missing a few very common `Collector` factories. To that effect, those are provided by
 the `ExtendedCollectors` class.
@@ -197,56 +183,6 @@ the `ExtendedCollectors` class.
 - `ExtendedCollectors.toArray` returns a collector that collects elements to an array.
 - `to[Primitive]Array` returns a collector that collects elements to a primitive array, where `[Primitive]` is the
   wanted primitive (e.g. `toIntArray`, `toDoubleArray`).
-
-## Kotlin Lambda Adapters for Java
-
-Kotlin's type system is very good, and the addition of the `Unit` type allows better type coherence in many cases.
-However, while in Java, one can very easily implement lambdas for Kotlin's function types for the most part, 
-if the function is supposed to return `Unit`, then the lambda will have to explicitly return the `Unit.INSTANCE`
-singleton, which is not convenient especially when the lambda is a one-liner, and when comparing it to Kotlin where it
-is the equivalent of return void.
-
-### LambdaAdapters
-
-The `LambdaAdapters` class provides a few static methods to adapt lambdas to Kotlin's function types. That is to say,
-they adapt the common Java functional interfaces which return `void` to Kotlin's function types which return `Unit`.
-(`Runnable`, `Consumer`, `BiConsumer`)
-
-It is important to note that the lambda is not wrapped in a Kotlin function type, it is actually a subtype of both the 
-Kotlin function type **and** the corresponding Java functional interface.
-
-```kt
-// Kotlin
-fun foo(block: (String) -> Unit) {
-    // ...
-}
-```
-
-```java
-// Java
-import static com.kamelia.sprinkler.util.jvmlambda.LambdaAdapters.*; // static import the `a` (adapt) functions
-
-class Main {
-    public static void main(String[] args) {
-        // you can write
-        foo(a(arg -> System.out.println(arg)));
-        
-        // instead of
-        foo(arg -> {
-            System.out.println(arg);
-            return Unit.INSTANCE;
-        });
-    }
-}
-```
-
-No need to explicitly return `Unit.INSTANCE` anymore.
-
-### InvokeExtensions
-
-This file provides an `invoke` extension operator to **every single** Java functional interface from the standard library.
-
-For example, on a `Consumer<T>`, it allows to call `consumer(value)` instead of `consumer.accept(value)`.
 
 ## ByteArrayDecoding
